fix(payouts): gate lot creation and campaign launch on Stripe Connect
Sellers and hub owners could create lots and start campaigns before
finishing Stripe Connect onboarding, leaving settlement with no payout
destination once buyers committed.

- /api/campaigns: refuse with 403 if hub owner has no Stripe Connect
  account, and 409 if the lot's seller has none (covers lots created
  before this gate existed).
- /api/lots/bulk: refuse with 403 if seller has no Stripe Connect
  account.
- Seller "Create Lot" page: server-side gate that renders a
  "Connect Stripe first" panel when the seller is not connected.
- Hub "Launch Campaign" page: same pattern for hub owners.

Existing campaigns test suite updated to mock the new profiles lookups,
plus two new tests covering the hub-owner and seller Stripe gates.
